Fixes made: fixed a potential seg fault in compute velocities, which is needed for VASP and vacf to work together
fixed the printing in read vasp so that the frames read is overwritten rather then having a new line

fixed in structure factor and fskt the error handling for method of k smapling.  the check for the value of k sampling was checked in compute rather then in check params, thus when i was using the functions I wasted all the time of reading the trajectory before being informed that there was an error in my input file.  So i moved this error check to check params that why the error is reported immediately.

fixed an issue that prevented alpha2 from being able to have a unique output file name.